Added document generator for creating random documents (driven by a dictionary file) whose tokens are zipf distributed. This will be the basis for a unified inverted-index testing framework.
git-svn-id: https://hyracks.googlecode.com/svn/branches/hyracks_inverted_index_updates_new@1816 123451ca-8445-de46-9d55-352943316053

+++ hyracks-test-support/src/main/java/edu/uci/ics/hyracks/storage/am/common/TestWorkloadConf.java
@@ -19,9 +19,9 @@ import edu.uci.ics.hyracks.storage.am.common.TestOperationSelector.TestOperation
 
 public class TestWorkloadConf {
     public final TestOperation[] ops;
-    public final float[] opProbs;
+    public final double[] opProbs;
 
-    public TestWorkloadConf(TestOperation[] ops, float[] opProbs) {
+    public TestWorkloadConf(TestOperation[] ops, double[] opProbs) {
         this.ops = ops;
         this.opProbs = opProbs;
     }
